requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
cbor2==5.8.0
celery==5.6.2
cffi==2.0.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.5
Django==6.0.2
django-cors-headers==4.9.0
django-otp==1.7.0
django-saintrelion-libs @ git+https://github.com/SaintRelion/Django-Fastly.git@8451943d31fc13d71d7a3ec6a41b5ce15026164e
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
kombu==5.6.2
packaging==26.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.41.0
vine==5.1.0
wcwidth==0.6.0
webauthn==2.7.1
